Add: Feature Tests
To simplify the testing approach a new TCP frontend for OSPD is written
in `ospd-tcp-frontend` with that it is possible to create a service
exposing the `:4242` so that the tests can be run without the hustle of
deployment definition.

In order to verify certain features of openvas a new project:
- feature-tests
is created.

The firs implementation will verify if the services of

- victim
- slsw

will be found.