Refactor: make it possible to add own paths (#3)
Refactor: make it possible to add own paths

To set your own paths instead of the defaults you have to follow the
instructions within `Use own paths instead of defaults` within the
README.md

In addition to ospd there is a new pod called openvas to display openvas
logs. It has everything mounted as ospd so there is the possibility to
play around in that in case the ospd container dies.