Change: make executing tests easier, reduce amount og release artifacts
Since the testing will need a Makefile eitherway a clone is still
necessary therefore the release will not contain any deployment
definitions anymore but only the build test binary.

In addition a new command to download the latest release is added and
the all command is defined to deploy, execute and delete the containers
after testing.